chore: split data-fetcher into separate process, add pm2
This does a lot of necessary cleanup. The most important things this commit does
is separates the "getting the data" logic from the web server itself. This means
requesting the app doesn't depend on ~11 other requests, and THIS means that the
app loads way faster!

data-fetcher.js is a new important file, which runs on a loop to update local data

A few unnecessary old ifles are deleted